Add Protected Endpoints
This commit adds an auth service, consumed as middleware, that checks
the request URL agains a list of protected URLs. If a match is found, it
ensures that a user (with UID) is also present on the request. The user
is put into the requst object by existing Passport strategies. If a user is
not found a 401 and message are returned; if a user is found, the next
piece of middleware is called.