fix(documents): replace constant with field path
I had incorrectly written this to use the contsant value of the model's title property instead of referencing the field when grouping. I believe this caused no major bugs because 1.) titles are unique in mEditor 2.) the previous step in the aggregation pipline sorted by last modfied. You can compare other `db.ts` files to see other examples of the correct usage of the `$group` stage.